Update dev menu with the new option
